Atualização Calculadora

+++ src/com/dio/calculadora/calculadora.java
@@ -8,17 +8,17 @@ public class calculadora {
 	{
 		Scanner scan = new Scanner(System.in);
 		
-		int a,b;
+		double a,b;
 		
 		System.out.println("Digite o primeiro valor: ");
 		a = scan.nextInt();
 		System.out.println("Digite o segundo valor: ");
 		b = scan.nextInt();
 		
-		int soma = soma(a,b);
-		int subtracao = subtracao(a,b);
-		int multiplicacao = multiplicacao(a,b);
-		int divisao = divisao(a,b);
+		double soma = soma(a,b);
+		double subtracao = subtracao(a,b);
+		double multiplicacao = multiplicacao(a,b);
+		double divisao = divisao(a,b);
 		
 		System.out.println(soma);
 		System.out.println(subtracao);
@@ -26,16 +26,16 @@ public class calculadora {
 		System.out.println(divisao);
 	}
 	
-	public static int soma(int a, int b) {
+	public static double soma(double a, double b) {
 		return a+b;
 	}
-	public static int subtracao(int a, int b) {
+	public static double subtracao(double a, double b) {
 		return a-b;
 	}
-	public static int multiplicacao(int a, int b) {
+	public static double multiplicacao(double a, double b) {
 		return a*b;
 	}
-	public static int divisao(int a, int b) {
+	public static double divisao(double a, double b) {
 		return a/b;
 	}
 }
